# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FASTANET ISP Billing & Hotspot Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all internet packages
      parameters:
        - name: orgId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Create a new package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageBody"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Update a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackageBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Delete a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Routers
...
  /pppoe-accounts:
    get:
      operationId: listPppoeAccounts
      tags: [pppoe]
      summary: List PPPoE accounts
      parameters:
        - name: orgId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of PPPoE accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PppoeAccount"
    post:
      operationId: createPppoeAccount
      tags: [pppoe]
      summary: Create a PPPoE account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePppoeAccountBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PppoeAccount"
  /pppoe-accounts/{id}:
    get:
      operationId: getPppoeAccount
      tags: [pppoe]
      summary: Get PPPoE account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PPPoE Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PppoeAccount"
    patch:
      operationId: updatePppoeAccount
      tags: [pppoe]
      summary: Update a PPPoE account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePppoeAccountBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PppoeAccount"
    delete:
      operationId: deletePppoeAccount
      tags: [pppoe]
      summary: Delete a PPPoE account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Payments
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: orgId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, paid, failed]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # Sessions
...
components:
  schemas:
...
    Package:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        name:
          type: string
        price:
          type: number
        durationValue:
          type: integer
        durationUnit:
          type: string
          enum: [minutes, hours, days]
        downloadSpeed:
          type: string
        uploadSpeed:
          type: string
        dataLimit:
          type: ["string", "null"]
        type:
          type: string
          enum: [hotspot, pppoe]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - name
        - price
        - durationValue
        - durationUnit
        - downloadSpeed
        - uploadSpeed
        - type
        - isActive
        - createdAt
    CreatePackageBody:
      type: object
      properties:
        orgId:
          type: integer
        name:
          type: string
        price:
          type: number
        durationValue:
          type: integer
        durationUnit:
          type: string
          enum: [minutes, hours, days]
        downloadSpeed:
          type: string
        uploadSpeed:
          type: string
        dataLimit:
          type: ["string", "null"]
        type:
          type: string
          enum: [hotspot, pppoe]
      required:
        - orgId
        - name
        - price
        - durationValue
        - durationUnit
        - downloadSpeed
        - uploadSpeed
        - type
    UpdatePackageBody:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        durationValue:
          type: integer
        durationUnit:
          type: string
          enum: [minutes, hours, days]
        downloadSpeed:
          type: string
        uploadSpeed:
          type: string
        dataLimit:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    PppoeAccount:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        routerId:
          type: integer
        packageId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        username:
          type: string
        service:
          type: string
        profile:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        expiryDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - routerId
        - username
        - service
        - status
        - createdAt
    CreatePppoeAccountBody:
      type: object
      properties:
        orgId:
          type: integer
        routerId:
          type: integer
        packageId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        username:
          type: string
        password:
          type: string
        service:
          type: string
        profile:
          type: ["string", "null"]
      required:
        - orgId
        - routerId
        - username
        - password
        - service
    UpdatePppoeAccountBody:
      type: object
      properties:
        packageId:
          type: ["integer", "null"]
        password:
          type: string
        service:
          type: string
        profile:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        expiryDate:
          type: ["string", "null"]
          format: date-time
    Payment:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        userId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        amount:
          type: number
        currency:
          type: string
        gateway:
          type: string
          enum: [mpesa, paystack, pesapal, intasend, manual]
        transactionRef:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, failed]
        userName:
          type: ["string", "null"]
        packageName:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - amount
        - currency
        - gateway
        - status
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        orgId:
          type: integer
        userId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        amount:
          type: number
        currency:
          type: string
        gateway:
          type: string
          enum: [mpesa, paystack, pesapal, intasend, manual]
        transactionRef:
          type: ["string", "null"]
        phoneNumber:
          type: ["string", "null"]
      required:
        - orgId
        - amount
        - currency
        - gateway
    UpdatePaymentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, failed]
        transactionRef:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
...